removed nfc scanner class
